Move config model types into a sub-namespace since some of the type names are quite common, thus, they can cause name conflict in the consumer's project
