+++ src/ConfigCatClient/ConfigCatClient.cs
@@ -9,6 +9,7 @@ using ConfigCat.Client.Cache;
 using ConfigCat.Client.ConfigService;
 using ConfigCat.Client.Configuration;
 using ConfigCat.Client.Evaluation;
+using ConfigCat.Client.Models;
 using ConfigCat.Client.Override;
 using ConfigCat.Client.Shims;
 using ConfigCat.Client.Utils;
