+++ src/ConfigCatClient/ConfigService/DefaultConfigFetcher.cs
@@ -5,6 +5,7 @@ using System.Net.Http.Headers;
 using System.Threading;
 using System.Threading.Tasks;
 using ConfigCat.Client.Configuration;
+using ConfigCat.Client.Models;
 using ConfigCat.Client.Shims;
 
 namespace ConfigCat.Client;
