+++ src/ConfigCatClient/Evaluation/EvaluateLogHelper.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Globalization;
+using ConfigCat.Client.Models;
 using ConfigCat.Client.Utils;
 
 namespace ConfigCat.Client.Evaluation;
