+++ src/ConfigCatClient/Evaluation/EvaluationDetails.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Runtime.CompilerServices;
 using ConfigCat.Client.Evaluation;
+using ConfigCat.Client.Models;
 using ConfigCat.Client.Utils;
 
 namespace ConfigCat.Client;
