+++ src/ConfigCatClient/Models/ConditionContainer.cs
@@ -1,7 +1,7 @@
 using System.Text.Json.Serialization;
 using ConfigCat.Client.Utils;
 
-namespace ConfigCat.Client;
+namespace ConfigCat.Client.Models;
 
 internal struct ConditionContainer : IConditionProvider
 {
