+++ src/ConfigCatClient/Models/PercentageOption.cs
@@ -2,7 +2,7 @@ using System.Text.Json.Serialization;
 using ConfigCat.Client.Evaluation;
 using ConfigCat.Client.Utils;
 
-namespace ConfigCat.Client;
+namespace ConfigCat.Client.Models;
 
 /// <summary>
 /// Describes a percentage option.
