+++ src/ConfigCatClient/Models/PrerequisiteFlagCondition.cs
@@ -2,7 +2,7 @@ using System.Text.Json.Serialization;
 using ConfigCat.Client.Evaluation;
 using ConfigCat.Client.Utils;
 
-namespace ConfigCat.Client;
+namespace ConfigCat.Client.Models;
 
 /// <summary>
 /// Describes a condition that is based on a prerequisite flag.
