+++ src/ConfigCatClient/Models/Segment.cs
@@ -4,7 +4,7 @@ using System.Text.Json.Serialization;
 using ConfigCat.Client.Evaluation;
 using ConfigCat.Client.Utils;
 
-namespace ConfigCat.Client;
+namespace ConfigCat.Client.Models;
 
 /// <summary>
 /// Describes a segment.
