+++ src/ConfigCatClient/Models/Setting.cs
@@ -5,7 +5,7 @@ using ConfigCat.Client.Evaluation;
 using ConfigCat.Client.Utils;
 using ConfigCat.Client.Versioning;
 
-namespace ConfigCat.Client;
+namespace ConfigCat.Client.Models;
 
 /// <summary>
 /// Describes a feature flag or setting.
