+++ src/ConfigCatClient/Models/SettingValue.cs
@@ -5,7 +5,7 @@ using System.Text.Json.Serialization;
 using ConfigCat.Client.Evaluation;
 using ConfigCat.Client.Utils;
 
-namespace ConfigCat.Client;
+namespace ConfigCat.Client.Models;
 
 internal struct SettingValue
 {
