+++ src/ConfigCatClient/Models/UserCondition.cs
@@ -4,7 +4,7 @@ using ConfigCat.Client.Evaluation;
 using ConfigCat.Client.Utils;
 using ConfigCat.Client.Versioning;
 
-namespace ConfigCat.Client;
+namespace ConfigCat.Client.Models;
 
 /// <summary>
 /// Describes a condition that is based on a User Object attribute.
