+++ src/ConfigCatClient/Override/IOverrideDataSource.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using ConfigCat.Client.Models;
 
 namespace ConfigCat.Client.Override;
