+++ src/ConfigCatClient/Override/LocalDictionaryDataSource.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using ConfigCat.Client.Models;
 
 namespace ConfigCat.Client.Override;
