+++ test/ConfigCat.Client.Tests/EvaluationTestsBase.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using ConfigCat.Client.Evaluation;
+using ConfigCat.Client.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace ConfigCat.Client.Tests;
