+++ test/ConfigCat.Client.Tests/Helpers/ConfigHelper.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
+using ConfigCat.Client.Models;
 
 namespace ConfigCat.Client.Tests.Helpers;
